+++ lib/test/src/test_min_cover.cpp
@@ -7,9 +7,9 @@ extern SimpleNetlist create_dwarf();        // import create_dwarf
 extern SimpleNetlist readNetD(const char *netDFileName);
 extern void readAre(SimpleNetlist &H, const char *areFileName);
 extern std::tuple<py::set<node_t>, int>
-min_net_cover_pd(SimpleNetlist &, const std::vector<size_t> &);
+min_net_cover_pd(SimpleNetlist &, const std::vector<int> &);
 extern std::tuple<py::set<node_t>, int>
-max_independent_net(SimpleNetlist &, const std::vector<size_t> &,
+max_independent_net(SimpleNetlist &, const std::vector<int> &,
                     const py::set<node_t> &);
 extern std::unique_ptr<SimpleNetlist>
   create_contraction_subgraph(SimpleNetlist &,
